atuaizando engine

+++ Vector2D/Window.h
@@ -2,7 +2,7 @@
 // Window (Arquivo de Cabeçalho)
 // 
 // Criação:     19 Mai 2007
-// Atualização: 23 Set 2021
+// Atualização: 01 Nov 2021
 // Compilador:  Visual C++ 2019
 //
 // Descrição:   A classe abstrai todos os detalhes de configuração de
@@ -57,8 +57,8 @@ public:
 
     HINSTANCE AppId();                                      // retorna o identificador da aplicação
     HWND Id();                                              // retorna o identificador da janela
-    int  Width();                                           // retorna a largura atual da janela
-    int  Height();                                          // retorna a altura atual da janela
+    float Width();                                          // retorna a largura atual da janela
+    float Height();                                         // retorna a altura atual da janela
 
     void Icon(const uint icon);                             // define o ícone da janela
     void Cursor(const uint cursor);                         // define o cursor da janela
@@ -102,12 +102,12 @@ inline HWND Window::Id()
 { return windowHandle; }
 
 // retorna a largura atual da janela
-inline int Window::Width() 
-{ return windowWidth;  }
+inline float Window::Width() 
+{ return float(windowWidth);  }
 
 // retorna a altura atual da janela
-inline int Window::Height() 
-{ return windowHeight; }
+inline float Window::Height() 
+{ return float(windowHeight); }
 
 // ----------------------------------------------------------
